requirements.txt
adbutils==2.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.14.3
boto3==1.42.41
botocore==1.42.41
cachetools==6.2.4
certifi==2025.7.14
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
Django==6.0.3
django-cors-headers==4.7.0
django-filter==25.2
django-jazzmin==3.0.3
django-storages==1.14.6
django-unfold==0.83.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
docxtpl==0.20.2
docx2pdf==0.1.8
drf-spectacular==0.28.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
faiss-cpu==1.13.1
fastapi==0.127.0
filelock==3.20.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.1.0
joblib==1.5.3
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==1.2.0
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.2.2
langchain-google-genai==4.1.2
langchain-openai==1.1.5
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.5.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
networkx==3.6.1
numpy==2.3.5
oauth2client==4.1.3
oauthlib==3.3.1
openai==2.13.0
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==12.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyGetWindow==0.0.9
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMySQL==1.1.1
pyparsing==3.2.5
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry2==0.9.5
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
sqlparse==0.5.3
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uiautomator2==3.5.0
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.12.0
uvicorn==0.40.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
